build dist for x86_64-sun-solaris
Presently, the standard library is built for "x86_64-sun-solaris"
targets.  We can and should build the full "dist" set of tools,
including rustc and cargo, so that they can be installed by Solaris and
illumos users via rustup.

This change creates a new Docker image, "dist-x86_64-sun-solaris", and
removes the x86 Solaris build from "dist-various-2".  It also adjusts
the manifest builder to expect complete toolchains for this platform.